check mail status in Keycloak
